feat (added tabs for graphs) - added Temperature, Humidity and Fill Level tab - styled analytics page
References Analytics of Data #49
