+++ package.json
@@ -8,7 +8,7 @@
     "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.8",
     "@mui/icons-material": "^5.14.18",
-    "@mui/material": "^5.14.18",
+    "@mui/material": "^5.14.20",
     "@mui/x-charts": "^6.18.3",
     "@tanstack/react-query": "^5.8.9",
     "@testing-library/jest-dom": "^5.17.0",
@@ -30,6 +30,7 @@
     "react-query": "^3.39.3",
     "react-router-dom": "^6.19.0",
     "react-scripts": "5.0.1",
+    "react-tabs": "^6.0.2",
     "string-to-color": "^2.2.2",
     "web-vitals": "^2.1.4"
   },
